+++ buildGui.sh
@@ -1,10 +1,11 @@
 #!/usr/bin/env bash
 
-if [ ! -d "tmp" ]; then
-    mkdir tmp
+version=$(git rev-list --all | head -n 1 | cut -b 1-5)
+if [ ! -e "gnc-2d-gui-v${version}.jar" ]; then
+    if [ ! -d "tmp" ]; then
+        mkdir tmp
+    fi
+    javac -encoding utf-8 -cp lib/javacsv.jar:lib/uiDesigner.jar -d tmp/ src/com/gensokyochess/*.java
+    jar cfm gnc-2d-gui-v${version}.jar META-INF/MANIFEST_GUI.MF -C tmp/ .
+    rm -rf tmp
 fi
-
-javac -encoding utf-8 -cp lib/javacsv.jar:lib/uiDesigner.jar -d tmp/ src/com/gensokyochess/*.java
-jar cfm gnc-2d-gui.jar META-INF/MANIFEST_GUI.MF -C tmp/ .
-
-rm -rf tmp
